+++ shenyu-client/shenyu-client-http/shenyu-client-springmvc/src/main/java/org/apache/shenyu/client/springmvc/register/apimeta/SpringMvcPreApiMetaBeanMatcher.java
@@ -15,21 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.client.springmvc.register;
+package org.apache.shenyu.client.springmvc.register.apimeta;
 
 import org.apache.shenyu.client.core.register.ApiBean;
 import org.apache.shenyu.client.core.register.matcher.Matcher;
 import org.apache.shenyu.client.springmvc.annotation.ShenyuSpringMvcClient;
 
-public class SpringMvcApiMetaBeanMatcher implements Matcher<ApiBean<Object>> {
+public class SpringMvcPreApiMetaBeanMatcher implements Matcher<ApiBean<Object>> {
 
     @Override
     public boolean match(final ApiBean<Object> element) {
         ShenyuSpringMvcClient annotation = element.getAnnotation(ShenyuSpringMvcClient.class);
-        if (annotation != null) {
-            return !annotation.path().endsWith("/**");
-        }
-        return true;
+        return annotation != null && annotation.path().endsWith("/**");
     }
-
 }
