+++ shenyu-client/shenyu-client-http/shenyu-client-springmvc/src/test/java/org/apache/shenyu/client/springmvc/register/apimeta/SpringMvcApiMetaBeanMatcherTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.client.springmvc.register;
+package org.apache.shenyu.client.springmvc.register.apimeta;
 
 import org.apache.shenyu.client.core.register.ApiBean;
 import org.apache.shenyu.client.springmvc.annotation.ShenyuSpringMvcClient;
@@ -24,31 +24,31 @@ import org.junit.jupiter.api.Test;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-public class SpringMvcPreApiMetaBeanMatcherTest {
+public class SpringMvcApiMetaBeanMatcherTest {
 
-    private final SpringMvcPreApiMetaBeanMatcher preApiMetaBeanMatcher = new SpringMvcPreApiMetaBeanMatcher();
+    private final SpringMvcApiMetaBeanMatcher beanMatcher = new SpringMvcApiMetaBeanMatcher();
 
     @Test
-    public void testBeanMatchWithAnnotation() throws Exception {
+    public void testBeanMatch() throws Exception {
         ApiBean<Object> apiBean = createSimpleApiBean(TestBeanMatchClass.class);
-        boolean result = preApiMetaBeanMatcher.match(apiBean);
-        assertThat(result, is(false));
+        boolean result = beanMatcher.match(apiBean);
+        assertThat(result, is(true));
     }
 
     @Test
     public void testBeanMatchWithoutAnnotation() throws Exception {
 
         ApiBean<Object> apiBean = createSimpleApiBean(TestBeanMatchWithoutAnnotationClass.class);
-        boolean result = preApiMetaBeanMatcher.match(apiBean);
-        assertThat(result, is(false));
+        boolean result = beanMatcher.match(apiBean);
+        assertThat(result, is(true));
     }
 
     @Test
     public void testBeanMatchWithStarAnnotation() throws Exception {
 
         ApiBean<Object> apiBean = createSimpleApiBean(TestBeanMatchWithStarAnnotationClass.class);
-        boolean result = preApiMetaBeanMatcher.match(apiBean);
-        assertThat(result, is(true));
+        boolean result = beanMatcher.match(apiBean);
+        assertThat(result, is(false));
     }
 
     private static ApiBean<Object> createSimpleApiBean(final Class<?> beanClass) throws Exception {
@@ -67,5 +67,4 @@ public class SpringMvcPreApiMetaBeanMatcherTest {
     @ShenyuSpringMvcClient("/**")
     static class TestBeanMatchWithStarAnnotationClass {
     }
-
 }
