+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-client/shenyu-spring-boot-starter-client-springmvc/src/main/java/org/apache/shenyu/springboot/starter/client/springmvc/ShenyuSpringMvcClientInfoRegisterConfiguration.java
@@ -27,12 +27,12 @@ import org.apache.shenyu.client.core.register.matcher.Matcher;
 import org.apache.shenyu.client.core.register.parser.ApiDocDefinitionParser;
 import org.apache.shenyu.client.core.register.parser.ApiMetaDefinitionParser;
 import org.apache.shenyu.client.core.register.parser.PreApiMetaBeanParser;
-import org.apache.shenyu.client.springmvc.register.SpringMvcApiBeansExtractor;
-import org.apache.shenyu.client.springmvc.register.SpringMvcApiDefinitionMetaMatcher;
-import org.apache.shenyu.client.springmvc.register.SpringMvcApiMetaBeanMatcher;
-import org.apache.shenyu.client.springmvc.register.SpringMvcApiMetaDefinitionParser;
-import org.apache.shenyu.client.springmvc.register.SpringMvcPreApiMetaBeanMatcher;
-import org.apache.shenyu.client.springmvc.register.SpringMvcPreApiMetaBeanParser;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcApiBeansExtractor;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcApiDefinitionMetaMatcher;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcApiMetaBeanMatcher;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcApiMetaDefinitionParser;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcPreApiMetaBeanMatcher;
+import org.apache.shenyu.client.springmvc.register.apimeta.SpringMvcPreApiMetaBeanParser;
 import org.apache.shenyu.common.enums.RpcTypeEnum;
 import org.apache.shenyu.common.utils.UriUtils;
 import org.apache.shenyu.register.common.config.PropertiesConfig;
